refactor: move registry inside detect dir
